add gitignore file.
