+++ .gitignore
@@ -1,0 +1,4 @@
+__pycache__/
+*.py[cod]
+.pytest_cache
+__pycache__/main.cpython-311.pyc
