+++ main.py
@@ -5,22 +5,7 @@ from flask import Flask
 
 # Create a Flask app
 app = Flask(__name__)
-def print_me(me: str):
-  print(me)
-
-# Define a route for the home page
-@app.route('/')
-def hello_world():
-    print("1 -Hello, World!")
-    job_that_executes_once()
-    print("2 - Hello, World!")
-    return 'Hello, World!'
-
-
-def job_that_executes_once():
-    # Do some work that only needs to happen once...
-    schedule.every(5).seconds.do(print_me, me="me!!!")
-    return schedule.CancelJob
+import views
 
 # Function to run the scheduler in a separate thread
 def run_scheduler():
